+++ src/index.js
@@ -1,0 +1,4 @@
+import BCDice_ from './BCDice';
+
+export const BCDice = BCDice_;
+export default BCDice;